Updated Dependencies
Changed the startup script to use OpenJDK-16 instead of the default-jre package in Ubuntu. Newer versions of Minecraft need this, and Ubuntu 20.04's default JRE is still OpenJDK-11. Additionally the project dependencies have been upgraded, which should fix some issues with the /status command and the latest versions of Minecraft.