Finished implementing the /reset command.
Added a cron job to the startup script that allows the minecraft server to start back up automatically after the droplet is rebooted by /reset. This is mainly useful for scenarios where the MC server crashes. Due to how DigitalOcean's billing works, if you run /stop and then /start, it is more expensive than /reset, because if you create and destroy a droplet they bill you for 1 hour minimum. Rebooting the droplet circumvents this.